refactor: simplify torrent management code
- Extract shared AbTorrentListPage component to deduplicate bangumi
  and orphan torrent pages (~240 lines of duplicate template+style)
- Remove unused ab-torrent-list.vue component (214 lines dead code)
- Fix useTorrentList hook: use project's auto-imported useMessage
  instead of naive-ui direct import, remove unused message export
- Optimize delete_orphans to use bulk DELETE SQL instead of
  select-all-then-delete-one-by-one
- Use .is_(None) instead of == None for SQLAlchemy NULL comparisons
- Add delete_obj method to avoid redundant double search in
  single-torrent delete API endpoints
- Replace all hardcoded Chinese strings with i18n keys
