+++ webui/src/i18n/en.json
@@ -192,6 +192,14 @@
       "downloaded": "Downloaded",
       "empty": "No torrents",
       "confirm_clear": "Are you sure you want to delete all torrents? This cannot be undone.",
+      "select_all": "Select All",
+      "deselect_all": "Deselect All",
+      "delete_selected": "Delete ({count})",
+      "deleted_one": "Deleted torrent {id}",
+      "deleted_count": "Deleted {count} torrents",
+      "cleared_all": "Cleared all torrents",
+      "load_failed": "Failed to load torrent list",
+      "delete_failed": "Failed to delete",
       "source": {
         "manual": "Manual"
       }
