+++ webui/src/i18n/zh-CN.json
@@ -192,6 +192,14 @@
       "downloaded": "已下载",
       "empty": "暂无种子",
       "confirm_clear": "确定要清空所有种子吗？此操作不可撤销。",
+      "select_all": "全选",
+      "deselect_all": "取消全选",
+      "delete_selected": "删除 ({count})",
+      "deleted_one": "已删除种子 {id}",
+      "deleted_count": "已删除 {count} 条种子",
+      "cleared_all": "已清空所有种子",
+      "load_failed": "加载种子列表失败",
+      "delete_failed": "删除失败",
       "source": {
         "manual": "手动/订阅"
       }
